Decompose CharacterCreator.tsx (c1): types, helpers, simple steps
First slice of the C-c CharacterCreator decomposition. Strict
mechanical extraction — zero behavior change, every helper /
component / type moves verbatim with imports adjusted.

New `apps/player-portal/src/routes/CharacterCreator/` directory:

- `types.ts` — wizard types (Step, PickerTarget, Slot, Draft,
  CreatorState, PickerFilters)
- `constants.ts` — EMPTY_DRAFT, STEPS, STEP_LABEL, PICKER_LABEL,
  STATIC_PICKER_FILTERS, BOOSTS_REQUIRED, ABILITY_LABEL
- `helpers.ts` — beginOrReusePendingActor / resetPendingActor,
  filtersForTarget, isStepFilled, featLocationFor,
  previousItemIdFor, persistPick, applyPickedSlot,
  prettySkillLabel, prettyLanguageLabel
- `CreatorSection.tsx` — section-shell wrapper used by every step
- `PickerCard.tsx` — selected-pick card used by Identity / Background
  / Ancestry / Class
- `FeatSlot.tsx` — compact L1 feat-slot chip used by Ancestry / Class
- `steps/IdentityStep.tsx` — identity text fields + deity picker
- `steps/AncestryStep.tsx` — ancestry / heritage / ancestry-feat
- `steps/ClassStep.tsx` — class + L1 features + class-feat
- `steps/ReviewStep.tsx` — final review summary

`CharacterCreator.tsx` (the route entry) shrinks to ~1100 LOC and
keeps: the root component, StepNav, AttributesStep + its 9 helpers
(BoostSourceBlock, FlawChips, BoostSlotPicker, FreeBoostBlock,
useSourceSlots, parseAncestryFlaws, parseAncestryOrBackgroundSlots,
parseClassKeyAbility, seedPicks), SkillsStep + helpers,
LanguagesStep + helpers, computeIntMod. These move in C-c2.

The `App.tsx` import (`from './routes/CharacterCreator'`) keeps
resolving to the .tsx file — TypeScript's bundler resolution prefers
the file over the sibling directory, so no consumer churn.

Bucket C-c1 of the cleanup audit. C-c2 (move the remaining steps +
extract a useCharacterDraft state hook) follows in its own PR.
