+++ apps/player-portal/src/routes/CharacterCreator.tsx
@@ -1,37 +1,38 @@
 import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { api } from '../api/client';
-import type { AbilityKey, CompendiumMatch, CompendiumSearchOptions } from '../api/types';
+import type { AbilityKey, CompendiumMatch } from '../api/types';
 import { ABILITY_KEYS } from '../api/types';
 import { BoostedMod } from '../components/creator/AbilityBoostPicker';
 import { FeatPicker } from '../components/creator/FeatPicker';
 import { PromptModal } from '../components/creator/PromptModal';
 import { usePendingPrompts } from '../lib/usePendingPrompts';
-import { useUuidHover } from '../lib/useUuidHover';
 
-// Module-scoped so React 18 StrictMode's dev-only double-mount
-// doesn't spawn two actor-create requests. First mount creates the
-// promise, second mount reuses it. Reset to null when the user
-// actually leaves the wizard (back or finish), so the next session
-// allocates a fresh draft actor.
-let pendingActorPromise: Promise<string> | null = null;
-function beginOrReusePendingActor(): Promise<string> {
-  if (pendingActorPromise === null) {
-    pendingActorPromise = api
-      .createActor({ name: 'New Character', type: 'character' })
-      .then((ref) => ref.id)
-      .catch((err: unknown) => {
-        // Clear so the next attempt retries fresh instead of getting
-        // stuck replaying the rejection.
-        pendingActorPromise = null;
-        throw err;
-      });
-  }
-  return pendingActorPromise;
-}
-function resetPendingActor(): void {
-  pendingActorPromise = null;
-}
+import { CreatorSection } from './CharacterCreator/CreatorSection';
+import {
+  ABILITY_LABEL,
+  BOOSTS_REQUIRED,
+  EMPTY_DRAFT,
+  PICKER_LABEL,
+  STEPS,
+  STEP_LABEL,
+} from './CharacterCreator/constants';
+import {
+  applyPickedSlot,
+  beginOrReusePendingActor,
+  filtersForTarget,
+  isStepFilled,
+  persistPick,
+  prettyLanguageLabel,
+  prettySkillLabel,
+  resetPendingActor,
+} from './CharacterCreator/helpers';
+import { PickerCard } from './CharacterCreator/PickerCard';
+import { AncestryStep } from './CharacterCreator/steps/AncestryStep';
+import { ClassStep } from './CharacterCreator/steps/ClassStep';
+import { IdentityStep } from './CharacterCreator/steps/IdentityStep';
+import { ReviewStep } from './CharacterCreator/steps/ReviewStep';
+import type { CreatorState, Draft, PickerTarget, Step } from './CharacterCreator/types';
 
 // Character creation wizard — Phase 1: identity + core choices.
 // Opening the wizard creates a blank actor in Foundry and the wizard
@@ -39,162 +40,6 @@ function resetPendingActor(): void {
 // step-advance; picks sync immediately (add the compendium item,
 // delete the previous pick for that slot). "Finish" lands the user
 // on the live sheet view for further allocation.
-
-type Step = 'identity' | 'ancestry' | 'class' | 'background' | 'attributes' | 'skills' | 'languages' | 'review';
-
-// Picker targets are decoupled from wizard steps: heritage selection
-// lives inside the ancestry step rather than owning a step of its own
-// (heritages are always children of an ancestry in pf2e's data), and
-// deity selection lives inside the identity step.
-type PickerTarget = 'ancestry' | 'heritage' | 'class' | 'background' | 'deity' | 'class-feat' | 'ancestry-feat';
-
-interface Slot {
-  match: CompendiumMatch;
-  // Item id on the persisted actor. Saved so we can delete the old
-  // item when the user changes their pick for this slot.
-  itemId: string;
-}
-
-interface Draft {
-  name: string;
-  // Free-text identity fields. pf2e stores these on `system.details`;
-  // we flush them to the actor when the user advances off the
-  // identity step (not per-keystroke).
-  gender: string;
-  age: string;
-  ethnicity: string;
-  nationality: string;
-  deity: Slot | null;
-  ancestry: Slot | null;
-  // Ancestry slug (e.g. 'elf', 'merfolk') fetched once after the
-  // ancestry is picked — heritage filtering needs it. Stored
-  // separately so an ancestry change can clear it while the refetch
-  // is in-flight.
-  ancestrySlug: string | null;
-  heritage: Slot | null;
-  // Heritage slug, fetched like ancestrySlug. Used so the ancestry-
-  // feat picker can surface versatile-heritage feats (changeling,
-  // aiuvarin, nephilim …) whose compendium items are tagged with the
-  // heritage slug rather than the parent ancestry's slug.
-  heritageSlug: string | null;
-  class: Slot | null;
-  // Class slug mirrors ancestrySlug — fetched from the class doc after
-  // a class is picked so the class-feat picker can scope to that
-  // trait. pf2e tags class feats with the class's slug as a trait
-  // (e.g. ['alchemist']).
-  classSlug: string | null;
-  background: Slot | null;
-  // Level-1 class and ancestry feat slots. Each is a regular picked
-  // item on the actor (same Slot shape as the other picks) — the
-  // creator just knows which pack/trait scope to search in.
-  classFeat: Slot | null;
-  ancestryFeat: Slot | null;
-  // Level-1 free boosts (four distinct abilities). Stored locally
-  // until the attributes step flushes them to
-  // `system.build.attributes.boosts.1` on the actor.
-  levelOneBoosts: AbilityKey[];
-  // User-selected skill trainings beyond the fixed ones baked into
-  // the ancestry/background/class items. pf2e applies everything in
-  // `class.system.trainedSkills.value` as rank-1 trainings, so the
-  // skills step appends these to that array (on top of what the
-  // class item ships with).
-  skillPicks: string[];
-  // Additional languages the user picked on top of the ancestry's
-  // fixed languages. Flushed into `actor.system.details.languages.value`
-  // merged with the granted list.
-  languagePicks: string[];
-  // Ambient state surfaced by the step components so the Review
-  // section can distinguish "user hasn't picked yet" from "no pick
-  // was ever available" (Wizard doesn't grant an L1 class feat;
-  // Anadi with Int 0 has no free language picks).
-  classGrantsL1Feat: boolean | null;
-  languageAllowance: number | null;
-  // Per-source boost picks. Each array mirrors the item's boost
-  // slots in order; fixed slots carry the pre-determined ability,
-  // choice/free slots start null until the user selects. Flushed to
-  // `system.build.attributes.boosts.{source}` on change.
-  ancestryBoosts: (AbilityKey | null)[];
-  backgroundBoosts: (AbilityKey | null)[];
-  classKeyAbility: AbilityKey | null;
-}
-
-const EMPTY_DRAFT: Draft = {
-  name: '',
-  gender: '',
-  age: '',
-  ethnicity: '',
-  nationality: '',
-  deity: null,
-  ancestry: null,
-  ancestrySlug: null,
-  heritage: null,
-  heritageSlug: null,
-  class: null,
-  classSlug: null,
-  background: null,
-  classFeat: null,
-  ancestryFeat: null,
-  levelOneBoosts: [],
-  ancestryBoosts: [],
-  backgroundBoosts: [],
-  classKeyAbility: null,
-  skillPicks: [],
-  languagePicks: [],
-  classGrantsL1Feat: null,
-  languageAllowance: null,
-};
-
-const STEPS: readonly Step[] = [
-  'identity',
-  'ancestry',
-  'class',
-  'background',
-  'attributes',
-  'skills',
-  'languages',
-  'review',
-];
-
-const STEP_LABEL: Record<Step, string> = {
-  identity: 'Identity',
-  ancestry: 'Ancestry',
-  class: 'Class',
-  background: 'Background',
-  attributes: 'Attributes',
-  skills: 'Skills',
-  languages: 'Languages',
-  review: 'Review',
-};
-
-const PICKER_LABEL: Record<PickerTarget, string> = {
-  ancestry: 'Ancestry',
-  heritage: 'Heritage',
-  class: 'Class',
-  background: 'Background',
-  deity: 'Deity',
-  'class-feat': 'Class Feat',
-  'ancestry-feat': 'Ancestry Feat',
-};
-
-type PickerFilters = Pick<
-  CompendiumSearchOptions,
-  'packIds' | 'documentType' | 'traits' | 'anyTraits' | 'ancestrySlug' | 'maxLevel'
->;
-
-const STATIC_PICKER_FILTERS: Record<
-  Exclude<PickerTarget, 'heritage' | 'class-feat' | 'ancestry-feat'>,
-  PickerFilters
-> = {
-  ancestry: { packIds: ['pf2e.ancestries'], documentType: 'Item' },
-  class: { packIds: ['pf2e.classes'], documentType: 'Item' },
-  background: { packIds: ['pf2e.backgrounds'], documentType: 'Item' },
-  deity: { packIds: ['pf2e.deities'], documentType: 'Item' },
-};
-
-// Actor lifecycle: wizard opens → creating → ready (actor exists in
-// Foundry, piecemeal patches flow through). Failed creation blocks
-// the UI with a retry button.
-type CreatorState = { kind: 'creating' } | { kind: 'ready'; actorId: string } | { kind: 'error'; message: string };
 
 export function CharacterCreator(): React.ReactElement {
   const navigate = useNavigate();
@@ -587,449 +432,6 @@ function StepNav({
         );
       })}
     </ol>
-  );
-}
-
-// Section shell used by the single-page creator layout. Each section
-// gets an anchor id (so the StepNav pills can scroll to it) and a
-// serif header matching the rest of the sheet. `scroll-mt` backs off
-// the sticky nav so a jumped-to section doesn't hide under it.
-function CreatorSection({
-  id,
-  title,
-  children,
-}: {
-  id: Step;
-  title: string;
-  children: React.ReactNode;
-}): React.ReactElement {
-  return (
-    <section
-      id={`creator-section-${id}`}
-      data-creator-section={id}
-      className="mb-6 scroll-mt-20 rounded border border-pf-border bg-white p-4"
-    >
-      <h2 className="mb-3 border-b border-pf-border pb-1 font-serif text-base font-semibold uppercase tracking-widest text-pf-alt-dark">
-        {title}
-      </h2>
-      {children}
-    </section>
-  );
-}
-
-// Identity text fields beyond name are free-form — pf2e stores them
-// as arbitrary strings on `system.details` and the sheet renders them
-// verbatim. Deity is the one exception; it has to land in the picker
-// because pf2e keys deity by compendium uuid for clergy/cleric gates
-// later on.
-type IdentityTextField = 'name' | 'gender' | 'age' | 'ethnicity' | 'nationality';
-
-function IdentityStep({
-  draft,
-  onChange,
-  onPickDeity,
-}: {
-  draft: Draft;
-  onChange: (patch: Partial<Draft>) => void;
-  onPickDeity: () => void;
-}): React.ReactElement {
-  const textFields: Array<{
-    key: IdentityTextField;
-    label: string;
-    placeholder: string;
-    autoFocus?: boolean;
-    fullWidth?: boolean;
-  }> = [
-    { key: 'name', label: 'Name', placeholder: 'e.g. Lutharion Saverin', autoFocus: true, fullWidth: true },
-    { key: 'gender', label: 'Gender / Pronouns', placeholder: 'e.g. she/her, non-binary' },
-    { key: 'age', label: 'Age', placeholder: 'e.g. 31' },
-    { key: 'ethnicity', label: 'Ethnicity', placeholder: 'e.g. Taldan' },
-    { key: 'nationality', label: 'Nationality', placeholder: 'e.g. Andoran' },
-  ];
-
-  return (
-    <div className="space-y-4">
-      <div className="grid grid-cols-1 gap-3 sm:grid-cols-2">
-        {textFields.map(({ key, label, placeholder, autoFocus, fullWidth }) => (
-          <label
-            key={key}
-            className={[
-              'block text-xs font-semibold uppercase tracking-widest text-pf-alt-dark',
-              fullWidth === true ? 'sm:col-span-2' : '',
-            ].join(' ')}
-          >
-            {label}
-            <input
-              id={`creator-${key}`}
-              type="text"
-              value={draft[key]}
-              onChange={(e): void => {
-                onChange({ [key]: e.target.value } as Partial<Draft>);
-              }}
-              autoFocus={autoFocus}
-              placeholder={placeholder}
-              className="mt-1 w-full rounded border border-pf-border bg-white px-3 py-2 text-sm font-normal normal-case tracking-normal text-pf-text focus:border-pf-primary focus:outline-none"
-            />
-          </label>
-        ))}
-      </div>
-      <div className="border-t border-pf-border pt-4" data-creator-subpicker="deity">
-        <PickerCard label="Deity" selection={draft.deity?.match ?? null} onOpen={onPickDeity} />
-      </div>
-    </div>
-  );
-}
-
-// After a class is picked we surface the auto-granted L1 features
-// pulled off `class.system.items` — the pf2e class chassis lists
-// each feature with its `{uuid, name, img, level}` and pf2e grants
-// everything `level <= characterLevel` when the class item is
-// attached. At creation we're always at level 1, so filter to that.
-// The list is read-only for now; the Progression tab handles the
-// per-level allocations once persistence is in place.
-interface ClassFeatureEntry {
-  uuid: string;
-  name: string;
-  img: string;
-  level: number;
-}
-
-type ClassDocState =
-  | { kind: 'idle' }
-  | { kind: 'loading'; uuid: string }
-  | {
-      kind: 'ready';
-      uuid: string;
-      features: ClassFeatureEntry[];
-      // pf2e's class chassis declares which levels grant a class
-      // feat; we use this to decide whether the L1 class-feat slot
-      // is valid (Wizard, Cleric, etc. don't grant L1 class feats).
-      grantsL1ClassFeat: boolean;
-    }
-  | { kind: 'error'; uuid: string; message: string };
-
-function ClassStep({
-  classPick,
-  classFeat,
-  classSlugResolved,
-  onPickClass,
-  onPickClassFeat,
-  onL1FeatAvailability,
-}: {
-  classPick: CompendiumMatch | null;
-  classFeat: CompendiumMatch | null;
-  classSlugResolved: boolean;
-  onPickClass: () => void;
-  onPickClassFeat: () => void;
-  // Bubbles the `classFeatLevels.value.includes(1)` result up to the
-  // parent once the class doc resolves, so the Review section can
-  // distinguish "slot vacant" from "class never granted a slot" for
-  // classes like Wizard or Cleric.
-  onL1FeatAvailability: (grants: boolean) => void;
-}): React.ReactElement {
-  const [docState, setDocState] = useState<ClassDocState>({ kind: 'idle' });
-  // Hover previews on the feature chips reuse the same stack-based
-  // popover plumbing as the rest of the app. Delegation handlers go
-  // on the features container; `data-uuid` on each chip triggers the
-  // fetch.
-  const uuidHover = useUuidHover();
-
-  useEffect(() => {
-    if (classPick === null) {
-      // eslint-disable-next-line react-hooks/set-state-in-effect
-      setDocState({ kind: 'idle' });
-      return;
-    }
-    const uuid = classPick.uuid;
-
-    setDocState({ kind: 'loading', uuid });
-    let cancelled = false;
-    void api
-      .getCompendiumDocument(uuid)
-      .then((res) => {
-        if (cancelled) return;
-        const features = extractLevel1Features(res.document.system);
-        const grantsL1ClassFeat = extractGrantsL1ClassFeat(res.document.system);
-        setDocState({ kind: 'ready', uuid, features, grantsL1ClassFeat });
-        onL1FeatAvailability(grantsL1ClassFeat);
-      })
-      .catch((err: unknown) => {
-        if (cancelled) return;
-        const message = err instanceof Error ? err.message : String(err);
-        setDocState({ kind: 'error', uuid, message });
-      });
-    return (): void => {
-      cancelled = true;
-    };
-  }, [classPick]);
-
-  return (
-    <div className="space-y-4">
-      <PickerCard label="Class" selection={classPick} onOpen={onPickClass} />
-      {classPick !== null && (
-        <div
-          className="space-y-4 border-t border-pf-border pt-4"
-          onMouseOver={uuidHover.delegationHandlers.onMouseOver}
-          onMouseOut={uuidHover.delegationHandlers.onMouseOut}
-        >
-          <div data-creator-subsection="class-features">
-            <h3 className="mb-2 font-serif text-xs font-semibold uppercase tracking-widest text-pf-alt-dark">
-              Level 1 Features
-            </h3>
-            <ClassFeaturesList state={docState} />
-          </div>
-          {docState.kind === 'ready' && docState.grantsL1ClassFeat && (
-            <div data-creator-subsection="class-feat">
-              <FeatSlot
-                label="Level 1 Class Feat"
-                selection={classFeat}
-                disabled={!classSlugResolved}
-                onOpen={onPickClassFeat}
-                {...(classSlugResolved ? {} : { disabledHint: 'Resolving class…' })}
-              />
-            </div>
-          )}
-          {docState.kind === 'ready' && !docState.grantsL1ClassFeat && (
-            <p className="text-xs italic text-pf-alt-dark" data-creator-subsection="class-feat-skip">
-              This class doesn&apos;t grant a class feat at level 1.
-            </p>
-          )}
-          {uuidHover.popover}
-        </div>
-      )}
-    </div>
-  );
-}
-
-// Compact feat-slot chip for the L1 class/ancestry feat picks on
-// the class/ancestry steps. Smaller and less imposing than the
-// full PickerCard used for ancestry/class/deity itself.
-function FeatSlot({
-  label,
-  selection,
-  disabled,
-  disabledHint,
-  onOpen,
-}: {
-  label: string;
-  selection: CompendiumMatch | null;
-  disabled?: boolean;
-  disabledHint?: string;
-  onOpen: () => void;
-}): React.ReactElement {
-  return (
-    <div className="flex items-center gap-2" data-feat-slot={label.toLowerCase().replace(/\s+/g, '-')}>
-      <span className="w-36 shrink-0 font-serif text-[11px] font-semibold uppercase tracking-widest text-pf-alt-dark">
-        {label}
-      </span>
-      {selection === null ? (
-        <>
-          <button
-            type="button"
-            onClick={onOpen}
-            disabled={disabled === true}
-            className="rounded border border-pf-border bg-white px-2 py-1 text-xs text-pf-text hover:bg-pf-bg-dark disabled:cursor-not-allowed disabled:opacity-50"
-          >
-            + Choose
-          </button>
-          {disabled === true && disabledHint !== undefined && (
-            <span className="text-xs italic text-pf-alt-dark">{disabledHint}</span>
-          )}
-        </>
-      ) : (
-        <>
-          <span
-            data-uuid={selection.uuid}
-            className="inline-flex items-center gap-1.5 rounded border border-pf-border bg-white px-2 py-1 text-xs text-pf-text"
-          >
-            <img src={selection.img} alt="" className="h-4 w-4 rounded bg-pf-bg-dark" />
-            <span className="truncate">{selection.name}</span>
-          </span>
-          <button
-            type="button"
-            onClick={onOpen}
-            className="rounded border border-pf-border bg-white px-2 py-1 text-[10px] uppercase tracking-widest text-pf-alt-dark hover:bg-pf-bg-dark"
-          >
-            Change
-          </button>
-        </>
-      )}
-    </div>
-  );
-}
-
-function ClassFeaturesList({ state }: { state: ClassDocState }): React.ReactElement {
-  if (state.kind === 'idle' || state.kind === 'loading') {
-    return <p className="text-xs italic text-pf-alt">Loading features…</p>;
-  }
-  if (state.kind === 'error') {
-    return <p className="text-xs text-pf-primary">Couldn&apos;t load class: {state.message}</p>;
-  }
-  if (state.features.length === 0) {
-    return <p className="text-xs italic text-pf-alt">No auto-granted features at level 1.</p>;
-  }
-  return (
-    <ul className="flex flex-wrap gap-2" data-testid="class-l1-features">
-      {state.features.map((f) => (
-        <li
-          key={f.uuid}
-          data-uuid={f.uuid}
-          className="inline-flex cursor-default items-center gap-1.5 rounded border border-pf-border bg-white px-2 py-1 text-xs text-pf-text"
-        >
-          <img src={f.img} alt="" className="h-4 w-4 rounded bg-pf-bg-dark" />
-          <span className="truncate">{f.name}</span>
-        </li>
-      ))}
-    </ul>
-  );
-}
-
-function extractGrantsL1ClassFeat(system: unknown): boolean {
-  const levels = (system as { classFeatLevels?: { value?: unknown } } | null)?.classFeatLevels?.value;
-  return Array.isArray(levels) && levels.some((v) => v === 1);
-}
-
-function extractLevel1Features(system: unknown): ClassFeatureEntry[] {
-  const items = (system as { items?: Record<string, unknown> } | null)?.items;
-  if (items === undefined) return [];
-  const out: ClassFeatureEntry[] = [];
-  for (const raw of Object.values(items)) {
-    if (typeof raw !== 'object' || raw === null) continue;
-    const entry = raw as { uuid?: unknown; name?: unknown; img?: unknown; level?: unknown };
-    if (typeof entry.uuid !== 'string' || typeof entry.name !== 'string') continue;
-    if (typeof entry.level !== 'number' || entry.level !== 1) continue;
-    out.push({
-      uuid: entry.uuid,
-      name: entry.name,
-      img: typeof entry.img === 'string' ? entry.img : '',
-      level: entry.level,
-    });
-  }
-  // Deterministic order: alphabetical so re-renders don't shuffle.
-  out.sort((a, b) => a.name.localeCompare(b.name));
-  return out;
-}
-
-function AncestryStep({
-  ancestry,
-  heritage,
-  ancestryFeat,
-  ancestrySlugResolved,
-  onPickAncestry,
-  onPickHeritage,
-  onPickAncestryFeat,
-}: {
-  ancestry: CompendiumMatch | null;
-  heritage: CompendiumMatch | null;
-  ancestryFeat: CompendiumMatch | null;
-  ancestrySlugResolved: boolean;
-  onPickAncestry: () => void;
-  onPickHeritage: () => void;
-  onPickAncestryFeat: () => void;
-}): React.ReactElement {
-  const uuidHover = useUuidHover();
-  return (
-    <div
-      className="space-y-4"
-      onMouseOver={uuidHover.delegationHandlers.onMouseOver}
-      onMouseOut={uuidHover.delegationHandlers.onMouseOut}
-    >
-      <PickerCard label="Ancestry" selection={ancestry} onOpen={onPickAncestry} />
-      {ancestry !== null && (
-        <div className="space-y-3 border-t border-pf-border pt-4">
-          <div data-creator-subpicker="heritage">
-            <PickerCard
-              label="Heritage"
-              selection={heritage}
-              onOpen={onPickHeritage}
-              disabled={!ancestrySlugResolved}
-              {...(ancestrySlugResolved ? {} : { disabledHint: 'Resolving ancestry…' })}
-            />
-          </div>
-          <div data-creator-subsection="ancestry-feat">
-            <FeatSlot
-              label="Level 1 Ancestry Feat"
-              selection={ancestryFeat}
-              disabled={!ancestrySlugResolved}
-              onOpen={onPickAncestryFeat}
-              {...(ancestrySlugResolved ? {} : { disabledHint: 'Resolving ancestry…' })}
-            />
-          </div>
-        </div>
-      )}
-      {uuidHover.popover}
-    </div>
-  );
-}
-
-function PickerCard({
-  label,
-  selection,
-  onOpen,
-  disabled,
-  disabledHint,
-}: {
-  label: string;
-  selection: CompendiumMatch | null;
-  onOpen: () => void;
-  disabled?: boolean;
-  disabledHint?: string;
-}): React.ReactElement {
-  if (selection === null) {
-    return (
-      <div className="flex flex-col items-start gap-2" data-picker-card={label.toLowerCase()}>
-        <p className="text-sm text-pf-text">No {label.toLowerCase()} selected yet.</p>
-        <button
-          type="button"
-          onClick={onOpen}
-          disabled={disabled === true}
-          className="rounded border border-pf-primary bg-pf-primary px-3 py-1.5 text-sm font-medium text-white hover:bg-pf-primary-dark disabled:cursor-not-allowed disabled:opacity-50"
-        >
-          Choose {label}
-        </button>
-        {disabled === true && disabledHint !== undefined && (
-          <p className="text-xs italic text-pf-alt-dark">{disabledHint}</p>
-        )}
-      </div>
-    );
-  }
-  return (
-    <div className="flex items-start gap-3" data-picker-card={label.toLowerCase()}>
-      {selection.img !== '' && (
-        <img
-          src={selection.img}
-          alt=""
-          className="h-14 w-14 flex-shrink-0 rounded border border-pf-border bg-pf-bg-dark"
-        />
-      )}
-      <div className="min-w-0 flex-1">
-        <p className="font-serif text-base font-semibold text-pf-text">{selection.name}</p>
-        <p className="text-[10px] uppercase tracking-widest text-pf-alt-dark">
-          {label}
-          {selection.level !== undefined && ` · Level ${selection.level.toString()}`}
-        </p>
-        {selection.traits !== undefined && selection.traits.length > 0 && (
-          <ul className="mt-1 flex flex-wrap gap-1">
-            {selection.traits.slice(0, 8).map((t) => (
-              <li
-                key={t}
-                className="rounded-full border border-pf-tertiary-dark bg-pf-tertiary/40 px-1.5 py-0.5 text-[10px] text-pf-alt-dark"
-              >
-                {t}
-              </li>
-            ))}
-          </ul>
-        )}
-      </div>
-      <button
-        type="button"
-        onClick={onOpen}
-        disabled={disabled === true}
-        className="rounded border border-pf-border bg-white px-2 py-1 text-xs text-pf-text hover:bg-pf-bg-dark disabled:cursor-not-allowed disabled:opacity-50"
-      >
-        Change
-      </button>
-    </div>
   );
 }
 
@@ -1926,13 +1328,6 @@ function normaliseAncestryLanguages(system: unknown): AncestryLanguagesDoc {
   return { fixed, suggested, bonusCount };
 }
 
-function prettyLanguageLabel(slug: string): string {
-  return slug
-    .split('-')
-    .map((w) => w.charAt(0).toUpperCase() + w.slice(1))
-    .join(' ');
-}
-
 function normaliseTrainedSkills(system: unknown): TrainedSkillsDoc {
   const ts = (system as { trainedSkills?: { value?: unknown; lore?: unknown; additional?: unknown } } | null)
     ?.trainedSkills;
@@ -1952,253 +1347,4 @@ function computeIntMod(draft: Draft): number {
   for (const k of draft.backgroundBoosts) if (k === 'int') n++;
   if (draft.classKeyAbility === 'int') n++;
   return n;
-}
-
-function prettySkillLabel(slug: string): string {
-  return slug
-    .split('-')
-    .map((w) => w.charAt(0).toUpperCase() + w.slice(1))
-    .join(' ');
-}
-
-// Sentinel used by ReviewStep so the renderer can tell "user hasn't
-// picked yet" from "there was nothing available to pick" (Wizard
-// class with no L1 class feat, Anadi with Int 0 → no extra
-// languages, etc.).
-const UNAVAILABLE = '__creator_unavailable__';
-
-function ReviewStep({ draft }: { draft: Draft }): React.ReactElement {
-  const textRow = (v: string): string | null => (v.trim().length > 0 ? v : null);
-  const rows: Array<[string, string | null]> = [
-    ['Name', textRow(draft.name)],
-    ['Gender / Pronouns', textRow(draft.gender)],
-    ['Age', textRow(draft.age)],
-    ['Ethnicity', textRow(draft.ethnicity)],
-    ['Nationality', textRow(draft.nationality)],
-    ['Deity', draft.deity?.match.name ?? null],
-    ['Ancestry', draft.ancestry?.match.name ?? null],
-    ['Heritage', draft.heritage?.match.name ?? null],
-    ['Ancestry Feat', draft.ancestryFeat?.match.name ?? null],
-    ['Class', draft.class?.match.name ?? null],
-    ['Class Feat', draft.classFeat?.match.name ?? (draft.classGrantsL1Feat === false ? UNAVAILABLE : null)],
-    ['Background', draft.background?.match.name ?? null],
-    ['L1 Boosts', draft.levelOneBoosts.length > 0 ? draft.levelOneBoosts.map((k) => k.toUpperCase()).join(', ') : null],
-    ['Free Skills', draft.skillPicks.length > 0 ? draft.skillPicks.map(prettySkillLabel).join(', ') : null],
-    [
-      'Additional Languages',
-      draft.languagePicks.length > 0
-        ? draft.languagePicks.map(prettyLanguageLabel).join(', ')
-        : draft.languageAllowance === 0
-          ? UNAVAILABLE
-          : null,
-    ],
-  ];
-  return (
-    <dl className="grid grid-cols-[auto_1fr] gap-x-4 gap-y-2 text-sm text-pf-text">
-      {rows.map(([label, value]) => (
-        <div key={label} className="contents">
-          <dt className="font-semibold uppercase tracking-widest text-pf-alt-dark">{label}</dt>
-          <dd>
-            {value === null ? (
-              <span className="italic text-neutral-400">Not chosen</span>
-            ) : value === UNAVAILABLE ? (
-              <span className="italic text-pf-alt-dark">Not granted by this character</span>
-            ) : (
-              value
-            )}
-          </dd>
-        </div>
-      ))}
-    </dl>
-  );
-}
-
-function filtersForTarget(target: PickerTarget, draft: Draft): PickerFilters {
-  if (target === 'heritage') {
-    const base: PickerFilters = { packIds: ['pf2e.heritages'], documentType: 'Item' };
-    if (draft.ancestrySlug !== null) base.ancestrySlug = draft.ancestrySlug;
-    return base;
-  }
-  if (target === 'class-feat') {
-    // pf2e tags class feats with the class's slug as a trait —
-    // 'alchemist', 'fighter', 'wizard' etc. Cap to level 1 so the
-    // picker only shows L1-qualifying feats.
-    const traits = draft.classSlug !== null ? [draft.classSlug] : undefined;
-    const base: PickerFilters = { packIds: ['pf2e.feats-srd'], documentType: 'Item', maxLevel: 1 };
-    if (traits) base.traits = traits;
-    return base;
-  }
-  if (target === 'ancestry-feat') {
-    // Pool the ancestry slug + heritage slug (when different) so the
-    // picker surfaces versatile-heritage feats (changeling, aiuvarin,
-    // nephilim …) alongside the parent ancestry's feats. When only
-    // the ancestry slug is known, fall back to a simple `traits`
-    // filter.
-    const base: PickerFilters = { packIds: ['pf2e.feats-srd'], documentType: 'Item', maxLevel: 1 };
-    const slugs: string[] = [];
-    if (draft.ancestrySlug !== null) slugs.push(draft.ancestrySlug);
-    if (draft.heritageSlug !== null && !slugs.includes(draft.heritageSlug)) slugs.push(draft.heritageSlug);
-    if (slugs.length > 1) {
-      base.anyTraits = slugs;
-    } else if (slugs.length === 1) {
-      base.traits = slugs;
-    }
-    return base;
-  }
-  return STATIC_PICKER_FILTERS[target];
-}
-
-function isStepFilled(step: Step, draft: Draft): boolean {
-  switch (step) {
-    case 'identity':
-      return draft.name.trim().length > 0;
-    case 'ancestry':
-      return draft.ancestry !== null && draft.heritage !== null;
-    case 'class':
-      return draft.class !== null;
-    case 'background':
-      return draft.background !== null;
-    case 'attributes':
-      return (
-        draft.levelOneBoosts.length === BOOSTS_REQUIRED &&
-        draft.ancestryBoosts.every((v) => v !== null) &&
-        draft.backgroundBoosts.every((v) => v !== null) &&
-        draft.classKeyAbility !== null
-      );
-    case 'skills':
-      return draft.skillPicks.length > 0;
-    case 'languages':
-      return draft.languagePicks.length > 0;
-    case 'review':
-      return false;
-  }
-}
-
-const BOOSTS_REQUIRED = 4;
-
-const ABILITY_LABEL: Record<AbilityKey, string> = {
-  str: 'Strength',
-  dex: 'Dexterity',
-  con: 'Constitution',
-  int: 'Intelligence',
-  wis: 'Wisdom',
-  cha: 'Charisma',
-};
-
-// Add the newly-picked compendium item to the actor, then delete the
-// previous pick for this slot (if any). Returns the new Slot for the
-// caller to commit into the draft. Order matters: add first so a
-// transient network failure doesn't leave the actor with zero items
-// for the slot.
-async function persistPick(actorId: string, target: PickerTarget, match: CompendiumMatch, draft: Draft): Promise<Slot> {
-  // L1 feats need a `system.location` slot tag so pf2e and the
-  // Progression tab recognise them as filling the slot. Non-feat
-  // picks don't need this — pf2e matches those via their category.
-  const location = featLocationFor(target);
-  const created = await api.addItemFromCompendium(actorId, {
-    packId: match.packId,
-    itemId: match.documentId,
-    ...(location !== null ? { systemOverrides: { location } } : {}),
-  });
-  const previousId = previousItemIdFor(draft, target);
-  if (previousId !== null) {
-    // Best-effort cleanup — if the old item vanished externally the
-    // delete returns 404 and that's fine for the user's state.
-    await api.deleteActorItem(actorId, previousId).catch(() => {
-      /* ignore */
-    });
-  }
-  // Heritage + ancestry feat get auto-discarded when ancestry changes —
-  // also clean up the old embedded items so the actor doesn't wear
-  // a dwarf heritage or an ancestry feat from the previous pick.
-  if (target === 'ancestry') {
-    const orphans = [draft.heritage?.itemId, draft.ancestryFeat?.itemId].filter(
-      (id): id is string => typeof id === 'string',
-    );
-    for (const id of orphans) {
-      await api.deleteActorItem(actorId, id).catch(() => {
-        /* ignore */
-      });
-    }
-  }
-  if (target === 'class' && draft.classFeat !== null) {
-    // Same rationale for class feat when the class changes.
-    await api.deleteActorItem(actorId, draft.classFeat.itemId).catch(() => {
-      /* ignore */
-    });
-  }
-  return { match, itemId: created.id };
-}
-
-// Feat-slot location strings mirror pf2e's own convention
-// (`<category>-<level>`). Only L1 feats at creation for now.
-function featLocationFor(target: PickerTarget): string | null {
-  if (target === 'class-feat') return 'class-1';
-  if (target === 'ancestry-feat') return 'ancestry-1';
-  return null;
-}
-
-function previousItemIdFor(draft: Draft, target: PickerTarget): string | null {
-  switch (target) {
-    case 'ancestry':
-      return draft.ancestry?.itemId ?? null;
-    case 'heritage':
-      return draft.heritage?.itemId ?? null;
-    case 'class':
-      return draft.class?.itemId ?? null;
-    case 'background':
-      return draft.background?.itemId ?? null;
-    case 'deity':
-      return draft.deity?.itemId ?? null;
-    case 'class-feat':
-      return draft.classFeat?.itemId ?? null;
-    case 'ancestry-feat':
-      return draft.ancestryFeat?.itemId ?? null;
-  }
-}
-
-function applyPickedSlot(draft: Draft, target: PickerTarget, slot: Slot): Draft {
-  switch (target) {
-    case 'ancestry':
-      // A new ancestry wipes the heritage + cached slug + ancestry
-      // feat + ancestry boost picks + language picks (the old
-      // choices may not be valid under the new ancestry).
-      return {
-        ...draft,
-        ancestry: slot,
-        ancestrySlug: null,
-        heritage: null,
-        heritageSlug: null,
-        ancestryFeat: null,
-        ancestryBoosts: [],
-        languagePicks: [],
-        languageAllowance: null,
-      };
-    case 'heritage':
-      // New heritage resets the cached slug + ancestry-feat pick
-      // (versatile heritages open up a different feat pool, so the
-      // previous pick may not still qualify).
-      return { ...draft, heritage: slot, heritageSlug: null, ancestryFeat: null };
-    case 'class':
-      // New class wipes the cached slug + class feat + key attribute
-      // pick for the same reason. Skill picks also reset since the
-      // free-skill count (class.additional) is class-specific.
-      return {
-        ...draft,
-        class: slot,
-        classSlug: null,
-        classFeat: null,
-        classKeyAbility: null,
-        skillPicks: [],
-        classGrantsL1Feat: null,
-      };
-    case 'background':
-      return { ...draft, background: slot, backgroundBoosts: [] };
-    case 'deity':
-      return { ...draft, deity: slot };
-    case 'class-feat':
-      return { ...draft, classFeat: slot };
-    case 'ancestry-feat':
-      return { ...draft, ancestryFeat: slot };
-  }
 }
